+++ packages/headless/src/controllers/breadcrumb-manager/headless-breadcrumb-manager.ts
@@ -18,7 +18,6 @@ import {
 } from '../../state/state-sections';
 import {BaseFacetRequest} from '../../features/facets/facet-api/request';
 import {executeSearch} from '../../features/search/search-actions';
-import {deselectAllFacets} from '../../features/facets/generic/facet-actions';
 import {logClearBreadcrumbs} from '../../features/facets/generic/facet-generic-analytics-actions';
 import {logFacetBreadcrumb} from '../../features/facets/facet-set/facet-set-analytics-actions';
 import {
@@ -41,6 +40,13 @@ import {
 } from '../../app/reducers';
 import {loadReducerError} from '../../utils/errors';
 import {SearchEngine} from '../../app/search-engine/search-engine';
+import {StaticFilterValue} from '..';
+import {StaticFilterSlice} from '../../features/static-filter-set/static-filter-set-state';
+import {
+  logStaticFilterDeselect,
+  toggleSelectStaticFilterValue,
+} from '../../features/static-filter-set/static-filter-set-actions';
+import {deselectAllBreadcrumbs} from '../../features/breadcrumb/breadcrumb-actions';
 
 /**
  * The `BreadcrumbManager` headless controller manages a summary of the currently active facet filters.
@@ -52,12 +58,10 @@ export interface BreadcrumbManager extends Controller {
   deselectAll(): void;
 
   /**
-   * Deselects a facet breadcrumb value or category facet breadcrumb.
-   * @param value - The facet breadcrumb value or category facet breadcrumb to deselect.
+   * Deselects a breadcrumb value.
+   * @param value - The breadcrumb value to deselect.
    */
-  deselectBreadcrumb(
-    value: BreadcrumbValue<BaseFacetValue> | CategoryFacetBreadcrumb
-  ): void;
+  deselectBreadcrumb(value: DeselectableValue): void;
 
   /**
    * The state of the `BreadcrumbManager` controller.
@@ -90,6 +94,11 @@ export interface BreadcrumbManagerState {
   dateFacetBreadcrumbs: DateFacetBreadcrumb[];
 
   /**
+   * The list of static filter breadcrumbs.
+   */
+  staticFilterBreadcrumbs: StaticFilterBreadcrumb[];
+
+  /**
    * Returns `true` if there are any available breadcrumbs (i.e., if there are any active facet values), and `false` if not.
    */
   hasBreadcrumbs: boolean;
@@ -113,7 +122,7 @@ export type DateFacetBreadcrumb = Breadcrumb<DateFacetValue>;
 /**
  * Represents a breadcrumb for a category facet.
  */
-export interface CategoryFacetBreadcrumb {
+export interface CategoryFacetBreadcrumb extends DeselectableValue {
   /**
    * The ID of the underlying facet.
    */
@@ -126,11 +135,27 @@ export interface CategoryFacetBreadcrumb {
    * The complete path to the underlying facet value.
    */
   path: CategoryFacetValue[];
-  /**
-   * Deselects the corresponding facet value.
-   */
-  deselect: () => void;
 }
+
+/**
+ * Represents a breadcrumb for a static filter.
+ */
+export interface StaticFilterBreadcrumb {
+  /**
+   * The ID of the underlying static filter.
+   */
+  id: string;
+
+  /**
+   * The list of static filter values currently selected.
+   */
+  values: StaticFilterBreadcrumbValue[];
+}
+
+/**
+ * Represents a static filter breadcrumb value.
+ */
+type StaticFilterBreadcrumbValue = BreadcrumbValue<StaticFilterValue>;
 
 /**
  * Represents a generic breadcrumb type.
@@ -155,15 +180,18 @@ export interface Breadcrumb<T extends BaseFacetValue> {
 /**
  * Represents a generic breadcrumb value type.
  */
-export interface BreadcrumbValue<T extends BaseFacetValue> {
+export interface BreadcrumbValue<T> extends DeselectableValue {
   /**
-   * The underlying facet value linked to this breadcrumb.
+   * The underlying value linked to this breadcrumb.
    */
   value: T;
+}
+
+export interface DeselectableValue {
   /**
-   * Deselects the corresponding facet value.
+   * A function that when called dispatches actions to deselect a breadcrumb value.
    */
-  deselect: () => void;
+  deselect(): void;
 }
 
 /**
@@ -276,12 +304,48 @@ export function buildBreadcrumbManager(
       .filter((breadcrumb) => breadcrumb.path.length);
   };
 
+  const getStaticFilterBreadcrumbs = (): StaticFilterBreadcrumb[] => {
+    const set = engine.state.staticFilterSet || {};
+    return Object.values(set).map(buildStaticFilterBreadcrumb);
+  };
+
+  const buildStaticFilterBreadcrumb = (
+    filter: StaticFilterSlice
+  ): StaticFilterBreadcrumb => {
+    const {id, values: filterValues} = filter;
+    const values = filterValues
+      .filter((value) => value.state === 'selected')
+      .map((value) => buildStaticFilterBreadcrumbValue(id, value));
+
+    return {id, values};
+  };
+
+  const buildStaticFilterBreadcrumbValue = (
+    id: string,
+    value: StaticFilterValue
+  ) => {
+    return {
+      value,
+      deselect: () => {
+        const {caption, expression} = value;
+        const analytics = logStaticFilterDeselect({
+          staticFilterId: id,
+          staticFilterValue: {caption, expression},
+        });
+
+        dispatch(toggleSelectStaticFilterValue({id, value}));
+        dispatch(executeSearch(analytics));
+      },
+    };
+  };
+
   function hasBreadcrumbs() {
     return !![
       ...getFacetBreadcrumbs(),
       ...getNumericFacetBreadcrumbs(),
       ...getDateFacetBreadcrumbs(),
       ...getCategoryFacetBreadcrumbs(),
+      ...getStaticFilterBreadcrumbs(),
     ].length;
   }
 
@@ -294,18 +358,17 @@ export function buildBreadcrumbManager(
         categoryFacetBreadcrumbs: getCategoryFacetBreadcrumbs(),
         numericFacetBreadcrumbs: getNumericFacetBreadcrumbs(),
         dateFacetBreadcrumbs: getDateFacetBreadcrumbs(),
+        staticFilterBreadcrumbs: getStaticFilterBreadcrumbs(),
         hasBreadcrumbs: hasBreadcrumbs(),
       };
     },
 
     deselectAll: () => {
-      dispatch(deselectAllFacets());
+      dispatch(deselectAllBreadcrumbs());
       dispatch(executeSearch(logClearBreadcrumbs()));
     },
 
-    deselectBreadcrumb(
-      value: BreadcrumbValue<BaseFacetValue> | CategoryFacetBreadcrumb
-    ) {
+    deselectBreadcrumb(value: DeselectableValue) {
       value.deselect();
     },
   };