+++ packages/headless/src/controllers/index.ts
@@ -217,10 +217,12 @@ export {
   FacetBreadcrumb,
   DateFacetBreadcrumb,
   CategoryFacetBreadcrumb,
+  StaticFilterBreadcrumb,
   Breadcrumb,
   BreadcrumbValue,
   BreadcrumbManagerState,
   BreadcrumbManager,
+  DeselectableValue,
   buildBreadcrumbManager,
 } from './breadcrumb-manager/headless-breadcrumb-manager';